Now at a form where we really just need to optimise the density and friction relations
Everything left to optimise is described in readme.txt

+++ davinci_1_upperboundary.f90
@@ -11,7 +11,7 @@ contains
 
   real function upperboundary(t)
 
-    integer, intent(in) :: t
+    real, intent(in) :: t
 
     upperboundary = 0.0
